Add user create details form. Move submit user logic to form. Remove client side password validation.
