+++ frontend/app/module/users/create/UserCreateContext.tsx
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 
-import { ApiError, ApiResult, CreateUserRequest, Role, User } from "@kiesraad/api";
+import type { Role } from "@kiesraad/api";
 
 export type UserType = "fullname" | "anonymous";
 
@@ -9,10 +9,6 @@ export interface IUserCreateContext {
   setRole: (role: Role) => void;
   type?: UserType;
   setType: (type: UserType) => void;
-  createUser: (user: CreateUserRequest) => Promise<ApiResult<User>>;
-  username?: string;
-  apiError: ApiError | null;
-  saving: boolean;
 }
 
 export const UserCreateContext = createContext<IUserCreateContext | undefined>(undefined);
