+++ frontend/lib/i18n/locales/nl/users.json
@@ -21,7 +21,7 @@
   "role_typist_hint": "Tellingen invoeren",
   "temporary_password": "Tijdelijk wachtwoord",
   "temporary_password_error_min_length": "Dit wachtwoord is niet lang genoeg. Gebruik minimaal {min_length} karakters",
-  "temporary_password_hint": "Gebruik minimaal {min_length} karakters. Het wachtwoord is hoofdlettergevoelig. De gebruiker moet na eerste keer inloggen zelf een nieuw wachtwoord kiezen.",
+  "temporary_password_hint": "Gebruik minimaal 13 karakters. Het wachtwoord is hoofdlettergevoelig. De gebruiker moet na eerste keer inloggen zelf een nieuw wachtwoord kiezen.",
   "type_anonymous": "Anonieme gebruikersnaam (bijvoorbeeld 'Invoerder01')",
   "type_fullname": "Op naam (bijvoorbeeld 'MariekeDeJager')",
   "type_hint": "Gebruikers met een anoniem account moeten bij de eerste keer inloggen hun echte naam invoeren. Anonieme accounts zijn vooral handig als je veel invoerders verwacht, en nog niet precies weet wie wel en wie niet aanwezig zal zijn.",
